+++ src/main/handlers/audioExtract.ts
@@ -1,5 +1,4 @@
 import path from 'path';
-
 import { path as ffmpegPath } from '@ffmpeg-installer/ffmpeg';
 import ffmpeg from 'fluent-ffmpeg';
