+++ src/main/handlers/index.ts
@@ -1,0 +1,11 @@
+import extractAudio_ from './audioExtract';
+import handleOpenProject_ from './openProjectHandler';
+import handleSaveProject_ from './saveProjectHandler';
+import handleTranscription_ from './transcriptionHandler';
+
+// This is kind of hacky so if you find a better way please change it lol
+
+export const extractAudio = extractAudio_;
+export const handleOpenProject = handleOpenProject_;
+export const handleSaveProject = handleSaveProject_;
+export const handleTranscription = handleTranscription_;
